Move CAS1 utils into service.cas1
